backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langgraph==0.2.50
langchain-core==0.3.21
langchain-anthropic==0.3.0
anthropic==0.40.0
sqlalchemy[asyncio]==2.0.35
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
alembic==1.14.0
redis==5.0.8
pydantic==2.9.0
email-validator>=2.0.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
geopy==2.4.1
shapely==2.0.6
firebase-admin==6.6.0
python-multipart==0.0.12
Pillow==11.0.0

# Optional (vector DB / workers) — install when needed:
# chromadb==0.5.20
# celery==5.4.0
